backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLite by default, SQLAlchemy ORM)
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM local inference (llama.cpp)
llama-cpp-python==0.2.80

# Optional: If you want to use Hugging Face locally, uncomment below
# transformers==4.35.2
# torch==2.1.1
# accelerate==0.24.1
# sentencepiece==0.1.99
# tokenizers==0.15.0

# HTTP client
httpx==0.25.2

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
